adding assignment #2 for c++ class
A company wants a program that calculates weekly gross pay, amount of deductions, and net pay for an hourly worker. The user will enter the number of hours worked for the week (allow partial hours) and the hourly pay rate. Use a named constant for a deduction rate of 27 percent. Output will be the two input values, the gross pay, deduction amount, and net pay. Desk-check the process with hours worked = 35 and pay rate = 12.75.